Moves everything config related to config.ts
